Added /contest-effect/ endpoint
