+++ src/api/components/contest-effect/contest-effect-controller.js
@@ -1,0 +1,18 @@
+const service = require('./contest-effect-service');
+
+// Example function to create an object and input it to the database
+async function get(request, response, next) {
+    try {
+        const { str } = request.params;
+
+        const dox = await service.getItem(str);
+
+        return response.status(200).json(dox);
+    } catch (error) {
+        return next(error);
+    }
+}
+
+module.exports = {
+    get,
+};
