+++ src/api/components/contest-effect/contest-effect-service.js
@@ -1,0 +1,8 @@
+const repository = require('./contest-effect-repository');
+
+async function getItem(name) {
+    return repository.getByID(name);
+}
+module.exports = {
+    getItem,
+};
